# src/TranscriberVCA.Generated/Controllers/AuthApi.cs
/*
 * TranscriberVCA
 *
 * Сервис транскрипции YouTube видео с помощью ИИ
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.Http;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using TranscriberVCA.Generated.Attributes;
using TranscriberVCA.Generated.Models;
using TranscriberVCA.Services;

namespace TranscriberVCA.Generated.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class AuthApiController : ControllerBase
    { 
        private readonly IAuthService _authService;
        
        public AuthApiController(IAuthService authService)
        {
            _authService = authService;
        }
        
        /// <summary>
        /// Авторизация
        /// </summary>
        /// <param name="loginRequest"></param>
        /// <response code="200">Успешная авторизация</response>
        /// <response code="401">Неверные учётные данные</response>
        [HttpPost]
        [Route("/auth/login")]
        [Consumes("application/json")]
        [ValidateModelState]
        [SwaggerOperation("Login")]
        [SwaggerResponse(statusCode: 200, type: typeof(AuthResponse), description: "Успешная авторизация")]
        [SwaggerResponse(statusCode: 401, type: typeof(Error), description: "Неверные учётные данные")]
        public virtual IActionResult Login([FromBody]LoginRequest loginRequest)
        {
            try
            {
                var response = _authService.Login(loginRequest);
                return Ok(response);
            }
            catch (UnauthorizedAccessException)
            {
                return Unauthorized(new Error { Detail = "Invalid credentials" });
            }
        }

        /// <summary>
        /// Регистрация нового пользователя
        /// </summary>
        /// <param name="registerRequest"></param>
        /// <response code="201">Пользователь создан</response>
        /// <response code="409">Пользователь уже существует</response>
        [HttpPost]
        [Route("/auth/register")]
        [Consumes("application/json")]
        [ValidateModelState]
        [SwaggerOperation("Register")]
        [SwaggerResponse(statusCode: 201, type: typeof(AuthResponse), description: "Пользователь создан")]
        [SwaggerResponse(statusCode: 409, type: typeof(Error), description: "Пользователь уже существует")]
        public virtual IActionResult Register([FromBody]RegisterRequest registerRequest)
        {
            try
            {
                var response = _authService.Register(registerRequest);
                return StatusCode(201, response);
            }
            catch (InvalidOperationException)
            {
                return StatusCode(409, new Error { Detail = "User already exists" });
            }
        }
    }
}