# src/TranscriberVCA.Generated/Controllers/HistoryApi.cs
/*
 * TranscriberVCA
 *
 * Сервис транскрипции YouTube видео с помощью ИИ
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.Http;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using TranscriberVCA.Generated.Attributes;
using TranscriberVCA.Generated.Models;
using TranscriberVCA.Services;

namespace TranscriberVCA.Generated.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class HistoryApiController : ControllerBase
    { 
        private readonly ITranscriptionService _transcriptionService;
        private readonly IAuthService _authService;

        public HistoryApiController(ITranscriptionService transcriptionService, IAuthService authService)
        {
            _transcriptionService = transcriptionService;
            _authService = authService;
        }
        /// <summary>
        /// История транскрипций пользователя
        /// </summary>
        /// <param name="limit">Количество записей</param>
        /// <param name="offset">Смещение</param>
        /// <response code="200">Список транскрипций пользователя</response>
        /// <response code="401">Не авторизован</response>
        [HttpGet]
        [Route("/history")]
        [Authorize]
        [ValidateModelState]
        [SwaggerOperation("GetHistory")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<HistoryItem>), description: "Список транскрипций пользователя")]
        [SwaggerResponse(statusCode: 401, type: typeof(Error), description: "Не авторизован")]
        public IActionResult GetHistory([FromQuery(Name = "limit")] int? limit, [FromQuery(Name = "offset")] int? offset)
        {
            var userId = _authService.ValidateToken(GetToken());
            if (userId == null)
                return Unauthorized(new Error { Detail = "Invalid token" });

            var result = _transcriptionService.GetHistory(userId, limit ?? 20, offset ?? 0);
            return Ok(result);
        }

        private string GetToken()
        {
            var authHeader = Request.Headers["Authorization"].ToString();
            return authHeader.Replace("Bearer ", "");
        }
    }
}