# src/TranscriberVCA.Generated/Controllers/TranscriptionsApi.cs
/*
 * TranscriberVCA
 *
 * Сервис транскрипции YouTube видео с помощью ИИ
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.Http;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using TranscriberVCA.Generated.Attributes;
using TranscriberVCA.Generated.Models;
using TranscriberVCA.Services;

namespace TranscriberVCA.Generated.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class TranscriptionsApiController : ControllerBase
    { 
        private readonly ITranscriptionService _transcriptionService;
        private readonly IAuthService _authService;

        public TranscriptionsApiController(ITranscriptionService transcriptionService, IAuthService authService)
        {
            _transcriptionService = transcriptionService;
            _authService = authService;
        }
        
        /// <summary>
        /// Получить результат транскрипции
        /// </summary>
        /// <param name="transcriptionId"></param>
        /// <response code="200">Результат транскрипции</response>
        /// <response code="404">Не найдено</response>
        /// <response code="401">Не авторизован</response>
        [HttpGet]
        [Route("/transcriptions/{transcription_id}")]
        [Authorize]
        [ValidateModelState]
        [SwaggerOperation("GetTranscription")]
        [SwaggerResponse(statusCode: 200, type: typeof(TranscriptionResult), description: "Результат транскрипции")]
        [SwaggerResponse(statusCode: 404, type: typeof(Error), description: "Не найдено")]
        [SwaggerResponse(statusCode: 401, type: typeof(Error), description: "Не авторизован")]
        public IActionResult GetTranscription([FromRoute(Name = "transcription_id")] [Required] string transcriptionId)
        {
            var userId = _authService.ValidateToken(GetToken());
            if (userId == null)
                return Unauthorized(new Error { Detail = "Invalid token" });

            var result = _transcriptionService.GetTranscription(userId, transcriptionId);
            if (result == null)
                return NotFound(new Error { Detail = "Not found" });

            return Ok(result);
        }

        /// <summary>
        /// Запросить транскрипцию YouTube видео
        /// </summary>
        /// <param name="transcribeRequest"></param>
        /// <response code="202">Транскрипция запущена</response>
        /// <response code="400">Невалидная ссылка</response>
        /// <response code="401">Не авторизован</response>
        [HttpPost]
        [Route("/transcribe")]
        [Authorize]
        [Consumes("application/json")]
        [ValidateModelState]
        [SwaggerOperation("Transcribe")]
        [SwaggerResponse(statusCode: 202, type: typeof(TranscriptionResult), description: "Транскрипция запущена")]
        [SwaggerResponse(statusCode: 400, type: typeof(Error), description: "Невалидная ссылка")]
        [SwaggerResponse(statusCode: 401, type: typeof(Error), description: "Не авторизован")]
        public IActionResult Transcribe([FromBody] TranscribeRequest transcribeRequest)
        {
            var userId = _authService.ValidateToken(GetToken());
            if (userId == null)
                return Unauthorized(new Error { Detail = "Invalid token" });

            var result = _transcriptionService.StartTranscription(userId, transcribeRequest);
            return StatusCode(202, result);
        }

        private string GetToken()
        {
            var authHeader = Request.Headers["Authorization"].ToString();
            return authHeader.Replace("Bearer ", "");
        }
    }
}